messaging_app/requirements.txt
asgiref==3.8.1
blinker==1.9.0
Bootstrap-Flask==2.4.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1

asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
django-environ==0.12.0
django-filter==25.1
django-silk==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gprof2dot==2025.4.14
mysqlclient==2.2.7
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2

